Add recommendations for the identifier format
These are additions to the general recommendations
for the MEF users and tools
to use and process identifiers in a portable way.

The recommendations are the same as proposed at the workshop
and taken from the experience with SCRAM.

To increase portability (especially, internationalization),
the suggestions are based on W3 XML schema datatype specifications.
The identifier specification is based on the XML NCName.

It is worth noting that the name specification prohibits "period"
because the MEF "references" specification relies on "period"
as a scope resolution operator.

The rest of suggestions are based on
how most modern XML libraries and tools can handle the XML input,
so it should not cause any issues
for modern software to provide these guarantees/conformance.