Transition from DTD to RelaxNG Compact schema
As presented at the workshop,
RelaxNG Compact (RNC) is a modern and convenient schema language
capable of fully describing features of the MEF
unlike the DTD.

The RNC syntax is very much the BNF-like readable,
leveraging regular expression operators.
The transition from the DTD to the RNC
should ease introduction of new MEF features.

The MEF schema in the RNC language is already provided
thanks to the early effort in the SCRAM project.
Only work necessary was to swap the schema in the Appendix B,
and update the wording in Chapter 2: Introduction.

The only current minor-drawback
is that syntax highlighter Pygments
doesn't yet provide support for RNC,
so it looks like a plain text or the EBNF.

Closes #11