Generalize Lognormal distribution specification
The formulas are generalized for various confidence levels.
The 95% confidence level formulas are given as an example.
A minor confusion of the expected value and the mean parameter
is fixed.

Unfortunately, there's not much discussion what the error factor means.
This could be expanded in future if needed.

Sources for the deduction of the general formula:

- Safety and Reliability: Proceedings of ESREL 2003, European Safety ..., Volume 2
- http://nrcoe.inl.gov/radscalc/Pages/AboutHTMLFiles/Distributions.htm

Fixes #21