Combine all layer schemas into major MEF schema
This schema combines all the schemas into the start grammar.
This major schema can be used for validation of input files.

Note that this major schema leverages RELAX NG include feature
instead of performing simple textual inclusion.
This allow a modular development of the schema
and provides conformance to SPOT principle.