Remove Appendix with the whole MEF schema
This schema can be in various formats
can be obtained from the schemas repository.
Since now the specification contains full schemas
(no more sketches or drafts in the formal specification!),
the specification text itself already contains the full schema
that can be cherry-picked by the implementers.

The Appendix is redundant.